requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
aiohttp==3.9.1
structlog==23.2.0
python-dotenv==1.0.0

firebase-admin==6.2.0
google-cloud-firestore==2.13.1